+++ src/modules/cars/repositories/implementations/CategoriesRepository.ts
@@ -1,5 +1,5 @@
-import { Category } from "../models/Category";
-import { ICategoriesRepository, ICreateCategoryDTO } from "./ICategoriesRepository";
+import { Category } from "../../models/Category";
+import { ICategoriesRepository, ICreateCategoryDTO } from "../ICategoriesRepository";
 
 class CategoriesRepository implements ICategoriesRepository {
