--- src/modules/cars/services/CreateSpecificationService.ts
@@ -1,28 +1,0 @@
-import { ISpecificationsRepository } from "../repositories/ISpecificationsRepository";
-
-interface IRequest {
-  name: string;
-  description: string;
-}
-class CreateSpecificationService {
-
-  constructor(private specificationsRepository: ISpecificationsRepository) {}
-
-  execute({ name, description }: IRequest): void {
-
-    const specificationAlreadyExists = this.specificationsRepository.findByName(name);
-
-    if (specificationAlreadyExists) {
-      throw new Error("Specification Already Exists!");
-    }
-
-    this.specificationsRepository.create({
-      name,
-      description,
-    });
-
-  }
-
-}
-
-export { CreateSpecificationService }
